add inital changelog entry if no lastTag and rename script
